requirements.txt
pytesseract==0.3.10
Pillow==9.5.0
opencv-python==4.7.0.72
numpy==1.24.3
pdf2image==1.16.3
PyQt6==6.5.0
pandas==2.0.1
openpyxl==3.1.2
PyMuPDF==1.22.5
fpdf==1.7.2
scikit-image==0.20.0
langdetect==1.0.9
loguru==0.7.0
pytest==7.3.1
qt-material==2.14